scaffold docker stuffsc
