+++ .env.production
@@ -1,0 +1,2 @@
+# the same db file should be configured for the other repos (jobber-classifier and jobber-site)
+DB_FILE=/path/to/db.sqlite
