+++ build.sh
@@ -1,0 +1,2 @@
+docker build -t nextjs-docker .
+# docker run -p 3000:3000 nextjs-docker
