+++ src/app/api/skills/route.ts
@@ -1,15 +1,4 @@
-import { db } from "@/database/db"
-
-export interface SkillDto {
-    id: number
-    name: string
-}
-
-export async function getSkills() {
-    const data = await db.selectFrom("skills").select(["id", "name"]).execute()
-
-    return data
-}
+import { getSkills } from "./handler"
 
 export async function GET() {
     return Response.json(await getSkills())
