+++ src/app/page.tsx
@@ -1,7 +1,7 @@
-import { getDuties } from "./api/duties/route"
-import { getJobs } from "./api/jobs/route"
-import { getLocations } from "./api/locations/route"
-import { getSkills } from "./api/skills/route"
+import { getDuties } from "./api/duties/handler"
+import { getJobs } from "./api/jobs/handler"
+import { getLocations } from "./api/locations/handler"
+import { getSkills } from "./api/skills/handler"
 import { HomeContainer } from "./home-container"
 import { deserializeParams } from "./search/useSearchForm"
 import { PageProps } from "./types"
