+++ src/app/search/search.tsx
@@ -3,9 +3,9 @@
 import { Button, Divider, Paper, TextField } from "@mui/material"
 import { FormEvent } from "react"
 import { useForm } from "react-hook-form"
-import { DutyDto } from "../api/duties/route"
-import { LocationDto } from "../api/locations/route"
-import { SkillDto } from "../api/skills/route"
+import { DutyDto } from "../api/duties/handler"
+import { LocationDto } from "../api/locations/handler"
+import { SkillDto } from "../api/skills/handler"
 import { useQueryParams } from "../useQueryParams"
 import { ClearanceFilter } from "./clearance-filter"
 import { DutyFilter } from "./duty-filter"
