+++ src/database/db.ts
@@ -2,7 +2,7 @@ import * as SQLite from "better-sqlite3"
 import { Kysely, SqliteDialect } from "kysely"
 import { Database } from "./types"
 
-const conn = new SQLite(process.env.DB_FILE)
+const conn = new (SQLite as any)(process.env.DB_FILE)
 conn.loadExtension("/usr/lib/sqlite3/pcre.so") // https://stackoverflow.com/a/8338515
 
 const dialect = new SqliteDialect({
